# src/swagger.json
{
  "basePath": "/",
  "consumes": ["application/json"],
  "info": {
    "title": "Passphrase Launchpad Server",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    }
  },
...
  "paths": {
...
    "/v1/investments/verify": {
      "post": {
        "tags": ["Investments"],
        "operationId": "verifySign",
        "summary": "Verify Sign",
        "description": "Endpoint for verify sign",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully verified",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                },
                "message": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "verifySign",
            "schema": {
              "properties": {
                "signatureId": {
                  "type": "string",
                  "default": null
                },
                "code": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["signatureId", "code"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/Investments/search": {
      "get": {
        "tags": ["Investments"],
        "operationId": "searchInvestments",
        "summary": "Search Investments",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Investments"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchInvestments",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "companyName",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
    "/v2/investments/verify": {
      "post": {
        "tags": ["Investments"],
        "operationId": "verifySign",
        "summary": "Verify Sign",
        "description": "Endpoint for verify sign",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully verified",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                },
                "message": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "verifySign",
            "schema": {
              "properties": {
                "signatureId": {
                  "type": "string",
                  "default": null
                },
                "code": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["signatureId", "code"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/Investments/search": {
      "get": {
        "tags": ["Investments"],
        "operationId": "searchInvestments",
        "summary": "Search Investments",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Investments"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchInvestments",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "companyName",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
    "/v1/investments/sign/{templateId}": {
      "post": {
        "tags": ["Investments"],
        "operationId": "signInvestment",
        "summary": "Sign Investment",
        "description": "Endpoint for sign investment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully signed",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "signatureId": {
                      "type": "string",
                      "default": null
                    },
                    "procedureId": {
                      "type": "string",
                      "default": null
                    },
                    "file": {
                      "type": "string",
                      "default": null
                    },
                    "fileId": {
                      "type": "string",
                      "default": null
                    }
                  }
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "properties": {
                "templateId": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["templateId"]
            }
          },
          {
            "in": "body",
            "name": "signInvestment",
            "schema": {
              "properties": {
                "amountInvested": {
                  "type": "number",
                  "default": null
                },
                "wallet": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object",
              "required": ["amountInvested", "wallet"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/investments/sign/{templateId}": {
      "post": {
        "tags": ["Investments"],
        "operationId": "signInvestment",
        "summary": "Sign Investment",
        "description": "Endpoint for sign investment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully signed",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "signatureId": {
                      "type": "string",
                      "default": null
                    },
                    "procedureId": {
                      "type": "string",
                      "default": null
                    },
                    "file": {
                      "type": "string",
                      "default": null
                    },
                    "fileId": {
                      "type": "string",
                      "default": null
                    }
                  }
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "properties": {
                "templateId": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["templateId"]
            }
          },
          {
            "in": "body",
            "name": "signInvestment",
            "schema": {
              "properties": {
                "amountInvested": {
                  "type": "number",
                  "default": null
                },
                "wallet": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object",
              "required": ["amountInvested", "wallet"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/investments/{companyId}": {
      "post": {
        "tags": ["Investments"],
        "operationId": "registerInvestment",
        "summary": "Register Investment",
        "description": "Endpoint for register investment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully registered",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "schema": {
              "properties": {
                "companyId": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["companyId"]
            }
          },
          {
            "in": "body",
            "name": "signInvestment",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/investments/{companyId}": {
      "post": {
        "tags": ["Investments"],
        "operationId": "registerInvestment",
        "summary": "Register Investment",
        "description": "Endpoint for register investment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully registered",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "schema": {
              "properties": {
                "companyId": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["companyId"]
            }
          },
          {
            "in": "body",
            "name": "signInvestment",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/investments/{id}/invest": {
      "put": {
        "tags": ["Investments"],
        "operationId": "makeAnInvestment",
        "summary": "Make An Investment",
        "description": "Endpoint for make an investment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully done",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "makeAnInvestment",
            "schema": {
              "properties": {
                "transactionHash": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/investments/{id}/invest": {
      "put": {
        "tags": ["Investments"],
        "operationId": "makeAnInvestment",
        "summary": "Make An Investment",
        "description": "Endpoint for make an investment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully done",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "makeAnInvestment",
            "schema": {
              "properties": {
                "transactionHash": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/investments/{id}/gasPrice": {
      "put": {
        "tags": ["Investments"],
        "operationId": "putInvestmentGas",
        "summary": "Put Investment Gas",
        "description": "Endpoint for put investment gas",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully done",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "putInvestmentGas",
            "schema": {
              "properties": {
                "gas": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/investments/{id}/gasPrice": {
      "put": {
        "tags": ["Investments"],
        "operationId": "putInvestmentGas",
        "summary": "Put Investment Gas",
        "description": "Endpoint for put investment gas",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully done",
            "schema": {
              "$ref": "#/definitions/Investments",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "putInvestmentGas",
            "schema": {
              "properties": {
                "gas": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/investments/amountSummary/{slug}": {
      "get": {
        "tags": ["Investments"],
        "operationId": "getInvestAmountSummary",
        "summary": "Get Invest Amount Summary",
        "description": "Endpoint for get invest amount summary",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                },
                "amount": {
                  "type": "number",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "schema": {
              "properties": {
                "slug": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["slug"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "produces": ["application/json"],
  "swagger": "2.0",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "body",
      "description": "This API requires authentication using a bearer token. Please include the token in the 'Authorization' header of your requests. ex: Bearer [space] <your-token>"
    }
  }
}